raise when the api returns a non successful response
This will somehow break our slack-notifier API. On successful
responses the flow will still be the same. In my projects I always
wrapped slack-notifier with a class which does excatly this. IMO when
there's an error with the slack API call shouldn't fail silently.

(I also added a changelog entry to indicate we should bump the version
from 2.1.0 to 2.2.0 to indicate that this could break the current code)