fix: attachments passed as a single hash should be wrapped in an array
at least at the time of this commit, slack does not support passing an
attachment as a hash, so while the link formatting for these types of
attachments has been fixed in the past - nothing was done to “array-ize”
them. this will do that. so if you pass a payload like:

  attachments: { text: “hello” }

it will be converted to:

  attachments: [{ text: “hello” }]