Update `RubyDoc.org` URL link in readme.md to Ruby version `3.0.1`
The URL link to `RubyDoc.org` in readme.md was a reference to
an older version of Ruby (`2.1` or `2.2`) that is no longer supported.

The document of Ruby 3.0 version already exists in `RubyDoc.org`,
and `slack-notifier` gem supports Ruby 3.0.

I think it's better to use the URL link to `RubyDoc.org` of version 3.0
in readme.md.